+++ dart/geobase/lib/vector_data.dart
@@ -20,8 +20,6 @@ library vector_data;
 
 export 'src/vector_data/array/coordinates.dart';
 export 'src/vector_data/array/coordinates_extensions.dart';
-export 'src/vector_data/coords/lonlat.dart';
-export 'src/vector_data/coords/xy.dart';
 export 'src/vector_data/model/bounded/bounded.dart';
 export 'src/vector_data/model/feature/feature.dart';
 export 'src/vector_data/model/feature/feature_builder.dart';