+++ dart/geobase/test/geodesy/spherical_test.dart
@@ -22,9 +22,6 @@ void main() {
     const p4 = Geographic(lat: 58.643889, lon: -3.07);
 
     test('Distance haversine', () {
-      // ignore: deprecated_member_use_from_same_package
-      expect(distanceHaversine(p3, p4), 968853.5441168448);
-
       expect(p1.spherical.distanceTo(p2), closeTo(404300, 300)); // 404.3×10³ m
       expect(p1.spherical.distanceTo(p2, radius: 3959),
           closeTo(251.2, 0.03)); // 251.2 mi